+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- configurer: merge `[common]` section into all actor group's sections.
+
 ## [0.1.23] - 2021-09-20
 ### Added
 - Graceful termination.